[REFACTOR][IR] Colocate type hooks and registrations
Group type-specific hook implementations and reflection registrations with their constructors. Move vector hooks into vector_expr.cc and organize structural, representation, and serialization registrations while preserving shared adapters and callback behavior.
