+++ src/ir/expr.cc
@@ -40,9 +40,11 @@
 
 namespace tvm {
 
-namespace {
+TVM_FFI_STATIC_INIT_BLOCK() { ExprNode::RegisterReflection(); }
 
-// Structural traversal hooks
+TVM_FFI_STATIC_INIT_BLOCK() { BaseFuncNode::RegisterReflection(); }
+
+namespace {
 
 TVMFFIAny OpaqueExprVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const OpaqueExprNode* self =
@@ -77,6 +79,20 @@ TVMFFIAny OpaqueExprMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   if (!mapped_ty.IsUnchanged()) self->ty = std::move(mapped_ty).ValueUnchecked();
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  OpaqueExprNode::RegisterReflection();
+  refl::TypeAttrDef<OpaqueExprNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&OpaqueExprVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&OpaqueExprMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&OpaqueExprMaybeInplaceMutate));
+}
+
+namespace {
 
 TVMFFIAny TensorLoadVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const TensorLoadNode* self =
@@ -127,6 +143,42 @@ TVMFFIAny TensorLoadMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
 
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  TensorLoadNode::RegisterReflection();
+  refl::TypeAttrDef<TensorLoadNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TensorLoadVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TensorLoadMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TensorLoadMaybeInplaceMutate));
+}
+
+// Tuple
+Tuple::Tuple(ffi::Array<Expr> fields, Span span) {
+  ffi::Optional<Type> tuple_ty = [&]() -> ffi::Optional<Type> {
+    ffi::Array<Type> field_ty;
+    for (const Expr& field : fields) {
+      if (field->ty.IsMissing()) {
+        return std::nullopt;
+      }
+      field_ty.push_back(field->ty);
+    }
+    return TupleType(field_ty);
+  }();
+
+  ffi::ObjectPtr<TupleNode> node = ffi::make_object<TupleNode>();
+  node->fields = std::move(fields);
+  node->span = std::move(span);
+  if (tuple_ty.has_value()) {
+    node->ty = tuple_ty.value();
+  }
+  data_ = std::move(node);
+}
+
+namespace {
+
 TVMFFIAny TupleVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const TupleNode* self =
       ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const TupleNode>(value);
@@ -165,6 +217,40 @@ TVMFFIAny TupleMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyVi
   if (!mapped_fields.IsUnchanged()) self->fields = std::move(mapped_fields).ValueUnchecked();
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  TupleNode::RegisterReflection();
+  refl::TypeAttrDef<TupleNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TupleVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TupleMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TupleMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.Tuple",
+                        [](ffi::Array<Expr> fields, Span span) { return Tuple(fields, span); });
+}
+
+// TupleGetItem
+TupleGetItem::TupleGetItem(Expr tuple, int index, Span span) {
+  TVM_FFI_CHECK_GE(index, 0, IndexError) << "Index out of bounds: Tuple " << tuple
+                                         << " cannot be accessed with negative index " << index;
+  ffi::ObjectPtr<TupleGetItemNode> node = ffi::make_object<TupleGetItemNode>();
+  if (const auto* tuple_type = tuple->ty.as<TupleTypeNode>()) {
+    TVM_FFI_CHECK_LT(index, tuple_type->fields.size(), IndexError)
+        << "Index out of bounds: Tuple " << tuple << " is of size " << tuple_type->fields.size()
+        << ", and cannot be accessed with index " << index;
+    node->ty = tuple_type->fields[index];
+  }
+  node->tuple = std::move(tuple);
+  node->index = index;
+  node->span = std::move(span);
+  data_ = std::move(node);
+}
+
+namespace {
 
 TVMFFIAny TupleGetItemVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   // skips: index
@@ -209,381 +295,20 @@ TVMFFIAny TupleGetItemMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
 
-TVMFFIAny IntImmVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny IntImmMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny IntImmMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny FloatImmVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny FloatImmMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny FloatImmMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  // skips: value
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny RangeVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
-  const RangeNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value);
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->min));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->extent));
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny RangeMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  const RangeNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value);
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_min,
-                                    mutator->MutateExpected(self->min));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_extent,
-                                    mutator->MutateExpected(self->extent));
-  if (mapped_min.UnchangedOrSameAs(self->min) && mapped_extent.UnchangedOrSameAs(self->extent)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  ffi::ObjectPtr<RangeNode> copy = ffi::make_object<RangeNode>(*self);
-  copy->min = std::move(mapped_min).ValueOrUnchanged(std::move(copy->min));
-  copy->extent = std::move(mapped_extent).ValueOrUnchanged(std::move(copy->extent));
-  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
-}
-
-TVMFFIAny RangeMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  RangeNode* self = const_cast<RangeNode*>(
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_min,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->min));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_extent,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->extent));
-  if (mapped_min.UnchangedOrSameAs(self->min) && mapped_extent.UnchangedOrSameAs(self->extent)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  if (!mapped_min.IsUnchanged()) self->min = std::move(mapped_min).ValueUnchecked();
-  if (!mapped_extent.IsUnchanged()) self->extent = std::move(mapped_extent).ValueUnchecked();
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-// DataflowVarNode duplicates this protocol because structural hooks do not inherit.  Keep the two
-// hook triples in lockstep when changing remap, PrimType-skip, or definition-region behavior.
-TVMFFIAny VarVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
-  // skips: name
-  const VarNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value);
-  // A PrimType carries only a dtype, so it has nothing to visit.  Broad callbacks do not see this
-  // skipped field; dynamically typed Vars still descend through the Type value.
-  if (!self->ty.as<PrimTypeNode>()) {
-    // Only Simple is clamped: Pattern co-introduces type fields such as BufferType shape
-    // variables, so that ambient region must continue through the dynamic type.
-    if (visitor->def_region_kind() == kTVMFFIDefRegionKindSimple) {
-      TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->WithDefRegionKind(
-          kTVMFFIDefRegionKindNone, [&]() { return visitor->VisitExpected(self->ty); }));
-    } else {
-      TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
-    }
-  }
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny VarMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  // skips: name
-  const VarNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value);
-  ffi::Expected<ffi::Any> remap_result = mutator->VarRemapGetExpected(value);
-  TVM_FFI_S_MUTATE_MAYBE_EARLY_RETURN(remap_result);
-  if (ffi::details::ExpectedUnsafe::GetData(remap_result).type_index() !=
-      ffi::TypeIndex::kTVMFFINone) {
-    return ffi::details::ExpectedUnsafe::MoveToTVMFFIAny(std::move(remap_result));
-  }
-  if (mutator->def_region_kind() == kTVMFFIDefRegionKindNone) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  ffi::UnchangedOr<ffi::Any> result = ffi::Unchanged();
-  ffi::Any mapped_value = ffi::Unchanged();
-  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
-  // this skipped field; dynamically typed Vars still descend through the Type value.
-  if (!self->ty.as<PrimTypeNode>()) {
-    // Pattern co-introduces type fields such as BufferType shape variables; Simple does not.
-    ffi::Expected<ffi::UnchangedOr<ffi::Any>> mapped_ty_result =
-        mutator->def_region_kind() == kTVMFFIDefRegionKindSimple
-            ? mutator->WithDefRegionKind(kTVMFFIDefRegionKindNone,
-                                         [&]() { return mutator->MutateExpected(self->ty); })
-            : mutator->MutateExpected(self->ty);
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
-                                      std::move(mapped_ty_result));
-    if (!mapped_ty.UnchangedOrSameAs(self->ty)) {
-      ffi::ObjectPtr<VarNode> copy = ffi::make_object<VarNode>(*self);
-      copy->ty = std::move(mapped_ty).ValueUnchecked();
-      mapped_value = ffi::Any(std::move(copy));
-      result = mapped_value;
-    }
-  }
-  if (!result.IsUnchanged() || mutator->def_region_kind() == kTVMFFIDefRegionKindPattern) {
-    auto set_result = mutator->VarRemapSetExpected(value, mapped_value);
-    if (TVM_FFI_PREDICT_FALSE(set_result.is_err())) {
-      return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(set_result).error()));
-    }
-  }
-  return ffi::details::UnchangedOrUnsafe::MoveToTVMFFIAny(std::move(result));
-}
-
-TVMFFIAny VarMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  // skips: name
-  VarNode* self = const_cast<VarNode*>(
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value));
-  ffi::Expected<ffi::Any> remap_result = mutator->VarRemapGetExpected(value);
-  TVM_FFI_S_MUTATE_MAYBE_EARLY_RETURN(remap_result);
-  if (ffi::details::ExpectedUnsafe::GetData(remap_result).type_index() !=
-      ffi::TypeIndex::kTVMFFINone) {
-    return ffi::details::ExpectedUnsafe::MoveToTVMFFIAny(std::move(remap_result));
-  }
-  if (mutator->def_region_kind() == kTVMFFIDefRegionKindNone) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  ffi::UnchangedOr<ffi::Any> result = ffi::Unchanged();
-  ffi::Any mapped_value = ffi::Unchanged();
-  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
-  // this skipped field; dynamically typed Vars still descend through the Type value.
-  if (!self->ty.as<PrimTypeNode>()) {
-    // Pattern co-introduces type fields such as BufferType shape variables; Simple does not.
-    ffi::Expected<ffi::UnchangedOr<ffi::Any>> mapped_ty_result =
-        mutator->def_region_kind() == kTVMFFIDefRegionKindSimple
-            ? mutator->WithDefRegionKind(
-                  kTVMFFIDefRegionKindNone,
-                  [&]() { return mutator->MaybeInplaceMutateIfUniqueExpected(self->ty); })
-            : mutator->MaybeInplaceMutateIfUniqueExpected(self->ty);
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
-                                      std::move(mapped_ty_result));
-    if (!mapped_ty.UnchangedOrSameAs(self->ty)) {
-      self->ty = std::move(mapped_ty).ValueUnchecked();
-      mapped_value = ffi::Any(self);
-      result = mapped_value;
-    }
-  }
-  if (!result.IsUnchanged() || mutator->def_region_kind() == kTVMFFIDefRegionKindPattern) {
-    auto set_result = mutator->VarRemapSetExpected(value, mapped_value);
-    if (TVM_FFI_PREDICT_FALSE(set_result.is_err())) {
-      return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(set_result).error()));
-    }
-  }
-  return ffi::details::UnchangedOrUnsafe::MoveToTVMFFIAny(std::move(result));
-}
-
-TVMFFIAny GlobalVarVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
-  // GlobalVar is a module-level symbol.  name_hint is scalar identity and ty is derived from the
-  // referenced function, matching GlobalVarNode's custom structural equality/hash definition.
-  // It has no definition site where this hook could establish a VarRemap.  A callback that renames
-  // GlobalVars is therefore responsible for returning one stable replacement per module symbol.
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny GlobalVarMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny GlobalVarMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-TVMFFIAny CallVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
-  // skips: attrs, constant metadata left untouched like the classic Expr functors.
-  const CallNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value);
-  // A PrimType carries only a dtype, so it has nothing to visit.  Broad callbacks do not see this
-  // skipped field; dynamically typed Call results still descend through the Type value.
-  if (!self->ty.as<PrimTypeNode>()) {
-    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
-  }
-  // An Op is an interned registry singleton, so it has nothing to visit.  Broad callbacks do not
-  // see this skipped field; function-valued Call operators still descend through the Expr value.
-  if (!self->op.as<OpNode>()) {
-    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->op));
-  }
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->args));
-  // An empty ty_args has no element to traverse.  Broad callbacks do not see the empty container;
-  // nonempty type arguments retain normal container descent and callback behavior.
-  if (!self->ty_args.empty()) {
-    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty_args));
-  }
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny CallMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  // skips: attrs, constant metadata left untouched like the classic Expr functors.
-  const CallNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value);
-  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
-  // this skipped field; dynamically typed Call results still descend through the Type value.
-  ffi::UnchangedOr<Type> mapped_ty = ffi::Unchanged();
-  if (!self->ty.as<PrimTypeNode>()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, descended_ty,
-                                      mutator->MutateExpected(self->ty));
-    mapped_ty = std::move(descended_ty);
-  }
-  // An Op is an interned registry singleton, so it has nothing to substitute.  Broad callbacks do
-  // not see this skipped field; function-valued Call operators still descend through the Expr.
-  ffi::UnchangedOr<Expr> mapped_op = ffi::Unchanged();
-  if (!self->op.as<OpNode>()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, descended_op,
-                                      mutator->MutateExpected(self->op));
-    mapped_op = std::move(descended_op);
-  }
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Expr>>, mapped_args,
-                                    mutator->MutateExpected(self->args));
-  // An empty ty_args has no element to substitute.  Broad callbacks do not see the empty
-  // container; nonempty type arguments retain normal container descent and callback behavior.
-  ffi::UnchangedOr<ffi::Array<Type>> mapped_ty_args = ffi::Unchanged();
-  if (!self->ty_args.empty()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Type>>, descended_ty_args,
-                                      mutator->MutateExpected(self->ty_args));
-    mapped_ty_args = std::move(descended_ty_args);
-  }
-  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_op.UnchangedOrSameAs(self->op) &&
-      mapped_args.UnchangedOrSameAs(self->args) &&
-      mapped_ty_args.UnchangedOrSameAs(self->ty_args)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  ffi::ObjectPtr<CallNode> copy = ffi::make_object<CallNode>(*self);
-  copy->ty = std::move(mapped_ty).ValueOrUnchanged(std::move(copy->ty));
-  copy->op = std::move(mapped_op).ValueOrUnchanged(std::move(copy->op));
-  copy->args = std::move(mapped_args).ValueOrUnchanged(std::move(copy->args));
-  copy->ty_args = std::move(mapped_ty_args).ValueOrUnchanged(std::move(copy->ty_args));
-  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
-}
-
-TVMFFIAny CallMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  // skips: attrs, constant metadata left untouched like the classic Expr functors.
-  CallNode* self = const_cast<CallNode*>(
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value));
-  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
-  // this skipped field; dynamically typed Call results still descend through the Type value.
-  ffi::UnchangedOr<Type> mapped_ty = ffi::Unchanged();
-  if (!self->ty.as<PrimTypeNode>()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, descended_ty,
-                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->ty));
-    mapped_ty = std::move(descended_ty);
-  }
-  // An Op is an interned registry singleton, so it has nothing to substitute.  Broad callbacks do
-  // not see this skipped field; function-valued Call operators still descend through the Expr.
-  ffi::UnchangedOr<Expr> mapped_op = ffi::Unchanged();
-  if (!self->op.as<OpNode>()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, descended_op,
-                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->op));
-    mapped_op = std::move(descended_op);
-  }
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Expr>>, mapped_args,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->args));
-  // An empty ty_args has no element to substitute.  Broad callbacks do not see the empty
-  // container; nonempty type arguments retain normal container descent and callback behavior.
-  ffi::UnchangedOr<ffi::Array<Type>> mapped_ty_args = ffi::Unchanged();
-  if (!self->ty_args.empty()) {
-    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Type>>, descended_ty_args,
-                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->ty_args));
-    mapped_ty_args = std::move(descended_ty_args);
-  }
-  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_op.UnchangedOrSameAs(self->op) &&
-      mapped_args.UnchangedOrSameAs(self->args) &&
-      mapped_ty_args.UnchangedOrSameAs(self->ty_args)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  if (!mapped_ty.IsUnchanged()) self->ty = std::move(mapped_ty).ValueUnchecked();
-  if (!mapped_op.IsUnchanged()) self->op = std::move(mapped_op).ValueUnchecked();
-  if (!mapped_args.IsUnchanged()) self->args = std::move(mapped_args).ValueUnchecked();
-  if (!mapped_ty_args.IsUnchanged()) self->ty_args = std::move(mapped_ty_args).ValueUnchecked();
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
 }  // namespace
 
-TVM_FFI_STATIC_INIT_BLOCK() { ExprNode::RegisterReflection(); }
-
-TVM_FFI_STATIC_INIT_BLOCK() { BaseFuncNode::RegisterReflection(); }
-
-// OpaqueExpr
-
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  OpaqueExprNode::RegisterReflection();
-  refl::TypeAttrDef<OpaqueExprNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&OpaqueExprVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&OpaqueExprMutate))
+  TupleGetItemNode::RegisterReflection();
+  refl::TypeAttrDef<TupleGetItemNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TupleGetItemVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TupleGetItemMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&OpaqueExprMaybeInplaceMutate));
-}
+            reinterpret_cast<void*>(&TupleGetItemMaybeInplaceMutate));
 
-// TensorLoad
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  TensorLoadNode::RegisterReflection();
-  refl::TypeAttrDef<TensorLoadNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TensorLoadVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TensorLoadMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TensorLoadMaybeInplaceMutate));
-}
-
-// Tuple
-Tuple::Tuple(ffi::Array<Expr> fields, Span span) {
-  ffi::Optional<Type> tuple_ty = [&]() -> ffi::Optional<Type> {
-    ffi::Array<Type> field_ty;
-    for (const Expr& field : fields) {
-      if (field->ty.IsMissing()) {
-        return std::nullopt;
-      }
-      field_ty.push_back(field->ty);
-    }
-    return TupleType(field_ty);
-  }();
-
-  ffi::ObjectPtr<TupleNode> node = ffi::make_object<TupleNode>();
-  node->fields = std::move(fields);
-  node->span = std::move(span);
-  if (tuple_ty.has_value()) {
-    node->ty = tuple_ty.value();
-  }
-  data_ = std::move(node);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  TupleNode::RegisterReflection();
-  refl::TypeAttrDef<TupleNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TupleVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TupleMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TupleMaybeInplaceMutate));
-}
-
-// TupleGetItem
-TupleGetItem::TupleGetItem(Expr tuple, int index, Span span) {
-  TVM_FFI_CHECK_GE(index, 0, IndexError) << "Index out of bounds: Tuple " << tuple
-                                         << " cannot be accessed with negative index " << index;
-  ffi::ObjectPtr<TupleGetItemNode> node = ffi::make_object<TupleGetItemNode>();
-  if (const auto* tuple_type = tuple->ty.as<TupleTypeNode>()) {
-    TVM_FFI_CHECK_LT(index, tuple_type->fields.size(), IndexError)
-        << "Index out of bounds: Tuple " << tuple << " is of size " << tuple_type->fields.size()
-        << ", and cannot be accessed with index " << index;
-    node->ty = tuple_type->fields[index];
-  }
-  node->tuple = std::move(tuple);
-  node->index = index;
-  node->span = std::move(span);
-  data_ = std::move(node);
+  refl::GlobalDef().def("ir.TupleGetItem", [](Expr tuple, int index, Span span) {
+    return TupleGetItem(tuple, index, span);
+  });
 }
 
 PrimExpr::PrimExpr(Call call) : PrimExpr(std::move(call).as_or_throw<PrimExpr>()) {}
@@ -612,19 +337,10 @@ PrimExpr TypeTraits<PrimExpr>::ConvertFallbackValue(double value) {
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  TupleGetItemNode::RegisterReflection();
   refl::GlobalDef()
-      .def("ir.Tuple", [](ffi::Array<Expr> fields, Span span) { return Tuple(fields, span); })
-      .def("ir.TupleGetItem",
-           [](Expr tuple, int index, Span span) { return TupleGetItem(tuple, index, span); })
       .def("relax.Tuple", [](ffi::Array<Expr> fields, Span span) { return Tuple(fields, span); })
       .def("relax.TupleGetItem",
            [](Expr tuple, int index, Span span) { return TupleGetItem(tuple, index, span); });
-  refl::TypeAttrDef<TupleGetItemNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&TupleGetItemVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&TupleGetItemMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TupleGetItemMaybeInplaceMutate));
 }
 
 // IntImm
@@ -662,17 +378,37 @@ IntImm::IntImm(PrimType value_ty, int64_t value, Span span) {
   data_ = std::move(node);
 }
 
+namespace {
+
+TVMFFIAny IntImmVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny IntImmMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+TVMFFIAny IntImmMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   IntImmNode::RegisterReflection();
-  refl::GlobalDef().def("ir.IntImm", [](DLDataType dtype, int64_t value, Span span) {
-    return IntImm(PrimType(dtype), value, span);
-  });
   refl::TypeAttrDef<IntImmNode>()
       .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&IntImmVisit))
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&IntImmMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&IntImmMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.IntImm", [](DLDataType dtype, int64_t value, Span span) {
+    return IntImm(PrimType(dtype), value, span);
+  });
 }
 
 // FloatImm
@@ -788,22 +524,104 @@ FloatImm::FloatImm(PrimType value_ty, double value, Span span) {
   data_ = std::move(node);
 }
 
+namespace {
+
+TVMFFIAny FloatImmVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny FloatImmMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+TVMFFIAny FloatImmMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  // skips: value
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   FloatImmNode::RegisterReflection();
-  refl::GlobalDef().def("ir.FloatImm", [](DLDataType dtype, double value, Span span) {
-    return FloatImm(PrimType(dtype), value, span);
-  });
   refl::TypeAttrDef<FloatImmNode>()
       .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&FloatImmVisit))
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&FloatImmMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&FloatImmMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.FloatImm", [](DLDataType dtype, double value, Span span) {
+    return FloatImm(PrimType(dtype), value, span);
+  });
 }
 
 // Range
 Range::Range(PrimExpr begin, PrimExpr end, Span span)
     : Range(ffi::make_object<RangeNode>(begin, tirx::is_zero(begin) ? end : (end - begin), span)) {}
+
+namespace {
+
+TVMFFIAny RangeVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
+  const RangeNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value);
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->min));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->extent));
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny RangeMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  const RangeNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value);
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_min,
+                                    mutator->MutateExpected(self->min));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_extent,
+                                    mutator->MutateExpected(self->extent));
+  if (mapped_min.UnchangedOrSameAs(self->min) && mapped_extent.UnchangedOrSameAs(self->extent)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::ObjectPtr<RangeNode> copy = ffi::make_object<RangeNode>(*self);
+  copy->min = std::move(mapped_min).ValueOrUnchanged(std::move(copy->min));
+  copy->extent = std::move(mapped_extent).ValueOrUnchanged(std::move(copy->extent));
+  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
+}
+
+TVMFFIAny RangeMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  RangeNode* self = const_cast<RangeNode*>(
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const RangeNode>(value));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_min,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->min));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<PrimExpr>, mapped_extent,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->extent));
+  if (mapped_min.UnchangedOrSameAs(self->min) && mapped_extent.UnchangedOrSameAs(self->extent)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  if (!mapped_min.IsUnchanged()) self->min = std::move(mapped_min).ValueUnchecked();
+  if (!mapped_extent.IsUnchanged()) self->extent = std::move(mapped_extent).ValueUnchecked();
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  RangeNode::RegisterReflection();
+  refl::TypeAttrDef<RangeNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&RangeVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&RangeMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&RangeMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.Range",
+                        [](PrimExpr begin, ffi::Optional<PrimExpr> end, Span span) -> Range {
+                          if (end.has_value()) {
+                            return Range(begin, end.value(), span);
+                          } else {
+                            return Range(IntImm(begin.ty(), 0), begin, span);
+                          }
+                        });
+}
 
 Range Range::FromMinExtent(PrimExpr min, PrimExpr extent, Span span) {
   return Range(ffi::make_object<RangeNode>(min, extent, span));
@@ -811,21 +629,7 @@ Range Range::FromMinExtent(PrimExpr min, PrimExpr extent, Span span) {
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  RangeNode::RegisterReflection();
-  refl::GlobalDef()
-      .def("ir.Range_from_min_extent", Range::FromMinExtent)
-      .def("ir.Range", [](PrimExpr begin, ffi::Optional<PrimExpr> end, Span span) -> Range {
-        if (end.has_value()) {
-          return Range(begin, end.value(), span);
-        } else {
-          return Range(IntImm(begin.ty(), 0), begin, span);
-        }
-      });
-  refl::TypeAttrDef<RangeNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&RangeVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&RangeMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&RangeMaybeInplaceMutate));
+  refl::GlobalDef().def("ir.Range_from_min_extent", Range::FromMinExtent);
 }
 
 // Var
@@ -837,6 +641,128 @@ Var::Var(ffi::String name, ffi::Optional<Type> ty_annotation, Span span) {
   }
   n->span = std::move(span);
   data_ = std::move(n);
+}
+
+namespace {
+
+// DataflowVarNode duplicates this protocol because structural hooks do not inherit.  Keep the two
+// hook triples in lockstep when changing remap, PrimType-skip, or definition-region behavior.
+TVMFFIAny VarVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
+  // skips: name
+  const VarNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value);
+  // A PrimType carries only a dtype, so it has nothing to visit.  Broad callbacks do not see this
+  // skipped field; dynamically typed Vars still descend through the Type value.
+  if (!self->ty.as<PrimTypeNode>()) {
+    // Only Simple is clamped: Pattern co-introduces type fields such as BufferType shape
+    // variables, so that ambient region must continue through the dynamic type.
+    if (visitor->def_region_kind() == kTVMFFIDefRegionKindSimple) {
+      TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->WithDefRegionKind(
+          kTVMFFIDefRegionKindNone, [&]() { return visitor->VisitExpected(self->ty); }));
+    } else {
+      TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
+    }
+  }
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny VarMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  // skips: name
+  const VarNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value);
+  ffi::Expected<ffi::Any> remap_result = mutator->VarRemapGetExpected(value);
+  TVM_FFI_S_MUTATE_MAYBE_EARLY_RETURN(remap_result);
+  if (ffi::details::ExpectedUnsafe::GetData(remap_result).type_index() !=
+      ffi::TypeIndex::kTVMFFINone) {
+    return ffi::details::ExpectedUnsafe::MoveToTVMFFIAny(std::move(remap_result));
+  }
+  if (mutator->def_region_kind() == kTVMFFIDefRegionKindNone) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::UnchangedOr<ffi::Any> result = ffi::Unchanged();
+  ffi::Any mapped_value = ffi::Unchanged();
+  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
+  // this skipped field; dynamically typed Vars still descend through the Type value.
+  if (!self->ty.as<PrimTypeNode>()) {
+    // Pattern co-introduces type fields such as BufferType shape variables; Simple does not.
+    ffi::Expected<ffi::UnchangedOr<ffi::Any>> mapped_ty_result =
+        mutator->def_region_kind() == kTVMFFIDefRegionKindSimple
+            ? mutator->WithDefRegionKind(kTVMFFIDefRegionKindNone,
+                                         [&]() { return mutator->MutateExpected(self->ty); })
+            : mutator->MutateExpected(self->ty);
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
+                                      std::move(mapped_ty_result));
+    if (!mapped_ty.UnchangedOrSameAs(self->ty)) {
+      ffi::ObjectPtr<VarNode> copy = ffi::make_object<VarNode>(*self);
+      copy->ty = std::move(mapped_ty).ValueUnchecked();
+      mapped_value = ffi::Any(std::move(copy));
+      result = mapped_value;
+    }
+  }
+  if (!result.IsUnchanged() || mutator->def_region_kind() == kTVMFFIDefRegionKindPattern) {
+    auto set_result = mutator->VarRemapSetExpected(value, mapped_value);
+    if (TVM_FFI_PREDICT_FALSE(set_result.is_err())) {
+      return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(set_result).error()));
+    }
+  }
+  return ffi::details::UnchangedOrUnsafe::MoveToTVMFFIAny(std::move(result));
+}
+
+TVMFFIAny VarMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  // skips: name
+  VarNode* self = const_cast<VarNode*>(
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const VarNode>(value));
+  ffi::Expected<ffi::Any> remap_result = mutator->VarRemapGetExpected(value);
+  TVM_FFI_S_MUTATE_MAYBE_EARLY_RETURN(remap_result);
+  if (ffi::details::ExpectedUnsafe::GetData(remap_result).type_index() !=
+      ffi::TypeIndex::kTVMFFINone) {
+    return ffi::details::ExpectedUnsafe::MoveToTVMFFIAny(std::move(remap_result));
+  }
+  if (mutator->def_region_kind() == kTVMFFIDefRegionKindNone) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::UnchangedOr<ffi::Any> result = ffi::Unchanged();
+  ffi::Any mapped_value = ffi::Unchanged();
+  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
+  // this skipped field; dynamically typed Vars still descend through the Type value.
+  if (!self->ty.as<PrimTypeNode>()) {
+    // Pattern co-introduces type fields such as BufferType shape variables; Simple does not.
+    ffi::Expected<ffi::UnchangedOr<ffi::Any>> mapped_ty_result =
+        mutator->def_region_kind() == kTVMFFIDefRegionKindSimple
+            ? mutator->WithDefRegionKind(
+                  kTVMFFIDefRegionKindNone,
+                  [&]() { return mutator->MaybeInplaceMutateIfUniqueExpected(self->ty); })
+            : mutator->MaybeInplaceMutateIfUniqueExpected(self->ty);
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
+                                      std::move(mapped_ty_result));
+    if (!mapped_ty.UnchangedOrSameAs(self->ty)) {
+      self->ty = std::move(mapped_ty).ValueUnchecked();
+      mapped_value = ffi::Any(self);
+      result = mapped_value;
+    }
+  }
+  if (!result.IsUnchanged() || mutator->def_region_kind() == kTVMFFIDefRegionKindPattern) {
+    auto set_result = mutator->VarRemapSetExpected(value, mapped_value);
+    if (TVM_FFI_PREDICT_FALSE(set_result.is_err())) {
+      return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(set_result).error()));
+    }
+  }
+  return ffi::details::UnchangedOrUnsafe::MoveToTVMFFIAny(std::move(result));
+}
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  VarNode::RegisterReflection();
+  refl::TypeAttrDef<VarNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&VarVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&VarMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&VarMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.Var", [](ffi::String name, ffi::Optional<Type> ty_annotation,
+                                     Span span) { return Var(name, ty_annotation, span); });
 }
 
 Var Var::CopyWithName(const ffi::String& name) const {
@@ -859,16 +785,6 @@ Var Var::CopyWithDType(PrimType dtype) const {
   return Var(std::move(copy));
 }
 
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  VarNode::RegisterReflection();
-  refl::TypeAttrDef<VarNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&VarVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&VarMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&VarMaybeInplaceMutate));
-}
-
 // GlobalVar
 GlobalVar::GlobalVar(ffi::String name_hint, Span span) {
   ffi::ObjectPtr<GlobalVarNode> n = ffi::make_object<GlobalVarNode>();
@@ -876,6 +792,26 @@ GlobalVar::GlobalVar(ffi::String name_hint, Span span) {
   n->span = std::move(span);
   data_ = std::move(n);
 }
+
+namespace {
+
+TVMFFIAny GlobalVarVisit(ffi::StructuralVisitorObj*, ffi::AnyView) noexcept {
+  // GlobalVar is a module-level symbol.  name_hint is scalar identity and ty is derived from the
+  // referenced function, matching GlobalVarNode's custom structural equality/hash definition.
+  // It has no definition site where this hook could establish a VarRemap.  A callback that renames
+  // GlobalVars is therefore responsible for returning one stable replacement per module symbol.
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny GlobalVarMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+TVMFFIAny GlobalVarMaybeInplaceMutate(ffi::StructuralMutatorObj*, ffi::AnyView) noexcept {
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
@@ -885,6 +821,8 @@ TVM_FFI_STATIC_INIT_BLOCK() {
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&GlobalVarMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&GlobalVarMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.GlobalVar", [](ffi::String name) { return GlobalVar(name); });
 }
 
 // Call
@@ -902,28 +840,142 @@ Call::Call(Type ret_ty, Expr op, ffi::Array<Expr> args, Attrs attrs, ffi::Array<
   data_ = std::move(n);
 }
 
+namespace {
+
+TVMFFIAny CallVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
+  // skips: attrs, constant metadata left untouched like the classic Expr functors.
+  const CallNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value);
+  // A PrimType carries only a dtype, so it has nothing to visit.  Broad callbacks do not see this
+  // skipped field; dynamically typed Call results still descend through the Type value.
+  if (!self->ty.as<PrimTypeNode>()) {
+    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
+  }
+  // An Op is an interned registry singleton, so it has nothing to visit.  Broad callbacks do not
+  // see this skipped field; function-valued Call operators still descend through the Expr value.
+  if (!self->op.as<OpNode>()) {
+    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->op));
+  }
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->args));
+  // An empty ty_args has no element to traverse.  Broad callbacks do not see the empty container;
+  // nonempty type arguments retain normal container descent and callback behavior.
+  if (!self->ty_args.empty()) {
+    TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty_args));
+  }
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny CallMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  // skips: attrs, constant metadata left untouched like the classic Expr functors.
+  const CallNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value);
+  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
+  // this skipped field; dynamically typed Call results still descend through the Type value.
+  ffi::UnchangedOr<Type> mapped_ty = ffi::Unchanged();
+  if (!self->ty.as<PrimTypeNode>()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, descended_ty,
+                                      mutator->MutateExpected(self->ty));
+    mapped_ty = std::move(descended_ty);
+  }
+  // An Op is an interned registry singleton, so it has nothing to substitute.  Broad callbacks do
+  // not see this skipped field; function-valued Call operators still descend through the Expr.
+  ffi::UnchangedOr<Expr> mapped_op = ffi::Unchanged();
+  if (!self->op.as<OpNode>()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, descended_op,
+                                      mutator->MutateExpected(self->op));
+    mapped_op = std::move(descended_op);
+  }
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Expr>>, mapped_args,
+                                    mutator->MutateExpected(self->args));
+  // An empty ty_args has no element to substitute.  Broad callbacks do not see the empty
+  // container; nonempty type arguments retain normal container descent and callback behavior.
+  ffi::UnchangedOr<ffi::Array<Type>> mapped_ty_args = ffi::Unchanged();
+  if (!self->ty_args.empty()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Type>>, descended_ty_args,
+                                      mutator->MutateExpected(self->ty_args));
+    mapped_ty_args = std::move(descended_ty_args);
+  }
+  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_op.UnchangedOrSameAs(self->op) &&
+      mapped_args.UnchangedOrSameAs(self->args) &&
+      mapped_ty_args.UnchangedOrSameAs(self->ty_args)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::ObjectPtr<CallNode> copy = ffi::make_object<CallNode>(*self);
+  copy->ty = std::move(mapped_ty).ValueOrUnchanged(std::move(copy->ty));
+  copy->op = std::move(mapped_op).ValueOrUnchanged(std::move(copy->op));
+  copy->args = std::move(mapped_args).ValueOrUnchanged(std::move(copy->args));
+  copy->ty_args = std::move(mapped_ty_args).ValueOrUnchanged(std::move(copy->ty_args));
+  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
+}
+
+TVMFFIAny CallMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  // skips: attrs, constant metadata left untouched like the classic Expr functors.
+  CallNode* self = const_cast<CallNode*>(
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const CallNode>(value));
+  // A PrimType carries only a dtype, so it has nothing to substitute.  Broad callbacks do not see
+  // this skipped field; dynamically typed Call results still descend through the Type value.
+  ffi::UnchangedOr<Type> mapped_ty = ffi::Unchanged();
+  if (!self->ty.as<PrimTypeNode>()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, descended_ty,
+                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->ty));
+    mapped_ty = std::move(descended_ty);
+  }
+  // An Op is an interned registry singleton, so it has nothing to substitute.  Broad callbacks do
+  // not see this skipped field; function-valued Call operators still descend through the Expr.
+  ffi::UnchangedOr<Expr> mapped_op = ffi::Unchanged();
+  if (!self->op.as<OpNode>()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, descended_op,
+                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->op));
+    mapped_op = std::move(descended_op);
+  }
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Expr>>, mapped_args,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->args));
+  // An empty ty_args has no element to substitute.  Broad callbacks do not see the empty
+  // container; nonempty type arguments retain normal container descent and callback behavior.
+  ffi::UnchangedOr<ffi::Array<Type>> mapped_ty_args = ffi::Unchanged();
+  if (!self->ty_args.empty()) {
+    TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<ffi::Array<Type>>, descended_ty_args,
+                                      mutator->MaybeInplaceMutateIfUniqueExpected(self->ty_args));
+    mapped_ty_args = std::move(descended_ty_args);
+  }
+  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_op.UnchangedOrSameAs(self->op) &&
+      mapped_args.UnchangedOrSameAs(self->args) &&
+      mapped_ty_args.UnchangedOrSameAs(self->ty_args)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  if (!mapped_ty.IsUnchanged()) self->ty = std::move(mapped_ty).ValueUnchecked();
+  if (!mapped_op.IsUnchanged()) self->op = std::move(mapped_op).ValueUnchecked();
+  if (!mapped_args.IsUnchanged()) self->args = std::move(mapped_args).ValueUnchecked();
+  if (!mapped_ty_args.IsUnchanged()) self->ty_args = std::move(mapped_ty_args).ValueUnchecked();
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   CallNode::RegisterReflection();
-  refl::GlobalDef()
-      .def("ir.Var", [](ffi::String name, ffi::Optional<Type> ty_annotation,
-                        Span span) { return Var(name, ty_annotation, span); })
-      .def("ir.GlobalVar", [](ffi::String name) { return GlobalVar(name); })
-      .def("ir.Call",
-           [](Type ret_ty, Expr op, ffi::Array<Expr> args, Attrs attrs, ffi::Array<Type> ty_args,
-              Span span) { return Call(ret_ty, op, args, attrs, ty_args, span); })
-      .def("ir.DebugPrint", [](ffi::ObjectRef ref) {
-        std::stringstream ss;
-        ss << ref;
-        return ss.str();
-      });
-  // Note: kRepr for GlobalVarNode is registered in script/printer/ir/ir.cc
-  // via TVM_REGISTER_SCRIPT_AS_REPR(GlobalVarNode, ReprPrintIR).
   refl::TypeAttrDef<CallNode>()
       .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&CallVisit))
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&CallMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&CallMaybeInplaceMutate));
+
+  refl::GlobalDef().def("ir.Call", [](Type ret_ty, Expr op, ffi::Array<Expr> args, Attrs attrs,
+                                      ffi::Array<Type> ty_args, Span span) {
+    return Call(ret_ty, op, args, attrs, ty_args, span);
+  });
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  refl::GlobalDef().def("ir.DebugPrint", [](ffi::ObjectRef ref) {
+    std::stringstream ss;
+    ss << ref;
+    return ss.str();
+  });
+  // Note: kRepr for GlobalVarNode is registered in script/printer/ir/ir.cc
+  // via TVM_REGISTER_SCRIPT_AS_REPR(GlobalVarNode, ReprPrintIR).
 }
 
 }  // namespace tvm
