+++ src/relax/distributed/type.cc
@@ -30,72 +30,9 @@ namespace tvm {
 namespace relax {
 namespace distributed {
 
-namespace {
-
-TVMFFIAny DTensorTypeVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
-  const DTensorTypeNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value);
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->device_mesh));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->placement));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->tensor_ty));
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
-}
-
-TVMFFIAny DTensorTypeMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  const DTensorTypeNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value);
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<DeviceMesh>, mapped_device_mesh,
-                                    mutator->MutateExpected(self->device_mesh));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Placement>, mapped_placement,
-                                    mutator->MutateExpected(self->placement));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<TensorType>, mapped_tensor_ty,
-                                    mutator->MutateExpected(self->tensor_ty));
-  if (mapped_device_mesh.UnchangedOrSameAs(self->device_mesh) &&
-      mapped_placement.UnchangedOrSameAs(self->placement) &&
-      mapped_tensor_ty.UnchangedOrSameAs(self->tensor_ty)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
-  }
-  ffi::ObjectPtr<DTensorTypeNode> copy = ffi::make_object<DTensorTypeNode>(*self);
-  copy->device_mesh = std::move(mapped_device_mesh).ValueOrUnchanged(std::move(copy->device_mesh));
-  copy->placement = std::move(mapped_placement).ValueOrUnchanged(std::move(copy->placement));
-  copy->tensor_ty = std::move(mapped_tensor_ty).ValueOrUnchanged(std::move(copy->tensor_ty));
-  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
-}
-
-TVMFFIAny DTensorTypeMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
-                                        ffi::AnyView value) noexcept {
-  DTensorTypeNode* self = const_cast<DTensorTypeNode*>(
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<DeviceMesh>, mapped_device_mesh,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->device_mesh));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Placement>, mapped_placement,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->placement));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<TensorType>, mapped_tensor_ty,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->tensor_ty));
-  if (!mapped_device_mesh.UnchangedOrSameAs(self->device_mesh)) {
-    self->device_mesh = std::move(mapped_device_mesh).ValueUnchecked();
-  }
-  if (!mapped_placement.UnchangedOrSameAs(self->placement)) {
-    self->placement = std::move(mapped_placement).ValueUnchecked();
-  }
-  if (!mapped_tensor_ty.UnchangedOrSameAs(self->tensor_ty)) {
-    self->tensor_ty = std::move(mapped_tensor_ty).ValueUnchecked();
-  }
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
-
-}  // namespace
-
 TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  DTensorTypeNode::RegisterReflection();
   PlacementNode::RegisterReflection();
   PlacementSpecNode::RegisterReflection();
-  refl::TypeAttrDef<DTensorTypeNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&DTensorTypeVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&DTensorTypeMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&DTensorTypeMaybeInplaceMutate));
 }
 
 PlacementSpec PlacementSpec::Sharding(int axis) {
@@ -199,6 +136,72 @@ DTensorType::DTensorType(TensorType tensor_ty, DeviceMesh device_mesh, Placement
       std::move(tensor_ty), std::move(device_mesh), std::move(placement));
   n->span = span;
   data_ = std::move(n);
+}
+
+namespace {
+
+TVMFFIAny DTensorTypeVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
+  const DTensorTypeNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value);
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->device_mesh));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->placement));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->tensor_ty));
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny DTensorTypeMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  const DTensorTypeNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value);
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<DeviceMesh>, mapped_device_mesh,
+                                    mutator->MutateExpected(self->device_mesh));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Placement>, mapped_placement,
+                                    mutator->MutateExpected(self->placement));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<TensorType>, mapped_tensor_ty,
+                                    mutator->MutateExpected(self->tensor_ty));
+  if (mapped_device_mesh.UnchangedOrSameAs(self->device_mesh) &&
+      mapped_placement.UnchangedOrSameAs(self->placement) &&
+      mapped_tensor_ty.UnchangedOrSameAs(self->tensor_ty)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::ObjectPtr<DTensorTypeNode> copy = ffi::make_object<DTensorTypeNode>(*self);
+  copy->device_mesh = std::move(mapped_device_mesh).ValueOrUnchanged(std::move(copy->device_mesh));
+  copy->placement = std::move(mapped_placement).ValueOrUnchanged(std::move(copy->placement));
+  copy->tensor_ty = std::move(mapped_tensor_ty).ValueOrUnchanged(std::move(copy->tensor_ty));
+  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
+}
+
+TVMFFIAny DTensorTypeMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
+                                        ffi::AnyView value) noexcept {
+  DTensorTypeNode* self = const_cast<DTensorTypeNode*>(
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DTensorTypeNode>(value));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<DeviceMesh>, mapped_device_mesh,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->device_mesh));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Placement>, mapped_placement,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->placement));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<TensorType>, mapped_tensor_ty,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->tensor_ty));
+  if (!mapped_device_mesh.UnchangedOrSameAs(self->device_mesh)) {
+    self->device_mesh = std::move(mapped_device_mesh).ValueUnchecked();
+  }
+  if (!mapped_placement.UnchangedOrSameAs(self->placement)) {
+    self->placement = std::move(mapped_placement).ValueUnchecked();
+  }
+  if (!mapped_tensor_ty.UnchangedOrSameAs(self->tensor_ty)) {
+    self->tensor_ty = std::move(mapped_tensor_ty).ValueUnchecked();
+  }
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  DTensorTypeNode::RegisterReflection();
+  refl::TypeAttrDef<DTensorTypeNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&DTensorTypeVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&DTensorTypeMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&DTensorTypeMaybeInplaceMutate));
 }
 
 TVM_FFI_STATIC_INIT_BLOCK() {
