+++ src/relax/ir/expr.cc
@@ -67,6 +67,121 @@ TVMFFIAny TypeOnlyExprMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
 
+// Hooks do not inherit, so DataflowVar must mirror the base VarNode remap, PrimType-skip, and
+// Simple-to-None definition-region protocol.  Keep this hook triple in lockstep with VarNode.
+
+// Parameters precede the reflected ty field in this hook triple so their pattern region establishes
+// the remap before the derived function type can refer to those symbolic definitions.
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() { BindingNode::RegisterReflection(); }
+
+If::If(Expr cond, Expr true_branch, Expr false_branch, Span span) {
+  ffi::ObjectPtr<IfNode> n = ffi::make_object<IfNode>();
+  n->cond = std::move(cond);
+  n->true_branch = std::move(true_branch);
+  n->false_branch = std::move(false_branch);
+  n->span = std::move(span);
+  data_ = std::move(n);
+}
+
+namespace {
+
+TVMFFIAny IfVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
+  const IfNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value);
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->cond));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->true_branch));
+  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->false_branch));
+  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}
+
+TVMFFIAny IfMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  const IfNode* self =
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value);
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
+                                    mutator->MutateExpected(self->ty));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, mapped_cond,
+                                    mutator->MutateExpected(self->cond));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_true_branch,
+                                    mutator->MutateExpected(self->true_branch));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_false_branch,
+                                    mutator->MutateExpected(self->false_branch));
+  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_cond.UnchangedOrSameAs(self->cond) &&
+      mapped_true_branch.UnchangedOrSameAs(self->true_branch) &&
+      mapped_false_branch.UnchangedOrSameAs(self->false_branch)) {
+    return ffi::Unchanged().CopyToTVMFFIAny();
+  }
+  ffi::ObjectPtr<IfNode> copy = ffi::make_object<IfNode>(*self);
+  copy->ty = std::move(mapped_ty).ValueOrUnchanged(std::move(copy->ty));
+  copy->cond = std::move(mapped_cond).ValueOrUnchanged(std::move(copy->cond));
+  copy->true_branch = std::move(mapped_true_branch).ValueOrUnchanged(std::move(copy->true_branch));
+  copy->false_branch =
+      std::move(mapped_false_branch).ValueOrUnchanged(std::move(copy->false_branch));
+  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
+}
+
+TVMFFIAny IfMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
+  IfNode* self = const_cast<IfNode*>(
+      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->ty));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, mapped_cond,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->cond));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_true_branch,
+                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->true_branch));
+  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(
+      ffi::UnchangedOr<SeqExpr>, mapped_false_branch,
+      mutator->MaybeInplaceMutateIfUniqueExpected(self->false_branch));
+  if (!mapped_ty.UnchangedOrSameAs(self->ty)) self->ty = std::move(mapped_ty).ValueUnchecked();
+  if (!mapped_cond.UnchangedOrSameAs(self->cond)) {
+    self->cond = std::move(mapped_cond).ValueUnchecked();
+  }
+  if (!mapped_true_branch.UnchangedOrSameAs(self->true_branch)) {
+    self->true_branch = std::move(mapped_true_branch).ValueUnchecked();
+  }
+  if (!mapped_false_branch.UnchangedOrSameAs(self->false_branch)) {
+    self->false_branch = std::move(mapped_false_branch).ValueUnchecked();
+  }
+  return ffi::Unchanged().CopyToTVMFFIAny();
+}
+
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  IfNode::RegisterReflection();
+  refl::TypeAttrDef<IfNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&IfVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&IfMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&IfMaybeInplaceMutate));
+
+  refl::GlobalDef().def("relax.If", [](Expr cond, Expr true_branch, Expr false_branch, Span span) {
+    return If(cond, true_branch, false_branch, span);
+  });
+}
+
+ShapeExpr::ShapeExpr(ffi::Array<PrimExpr> values, Span span) {
+  ffi::ObjectPtr<ShapeExprNode> n = ffi::make_object<ShapeExprNode>();
+
+  n->values = values.Map([](PrimExpr value) {
+    if (value->IsInstance<IntImmNode>()) {
+      return tvm::cast(PrimType::Int(64), value);
+    }
+    TVM_FFI_ICHECK(value.ty().MatchesElementType(DLDataTypeCode::kDLInt, 64))
+        << "the value in ShapeType can only have dtype of int64";
+    return value;
+  });
+  n->span = span;
+  n->ty = ShapeType(values, span);
+  data_ = std::move(n);
+}
+
+namespace {
+
 TVMFFIAny ShapeExprVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const ShapeExprNode* self =
       ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const ShapeExprNode>(value);
@@ -106,8 +221,34 @@ TVMFFIAny ShapeExprMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
 
-// Hooks do not inherit, so DataflowVar must mirror the base VarNode remap, PrimType-skip, and
-// Simple-to-None definition-region protocol.  Keep this hook triple in lockstep with VarNode.
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  ShapeExprNode::RegisterReflection();
+  refl::TypeAttrDef<ShapeExprNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&ShapeExprVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&ShapeExprMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&ShapeExprMaybeInplaceMutate));
+
+  refl::GlobalDef().def("relax.ShapeExpr", [](ffi::Array<PrimExpr> values, Span span) {
+    return ShapeExpr(values, span);
+  });
+}
+
+DataflowVar::DataflowVar(ffi::String name, ffi::Optional<Type> ty_annotation, Span span) {
+  ffi::ObjectPtr<DataflowVarNode> n = ffi::make_object<DataflowVarNode>();
+  n->name = std::move(name);
+  if (ty_annotation.has_value()) {
+    n->ty = ty_annotation.value();
+  }
+  n->span = std::move(span);
+  data_ = std::move(n);
+}
+
+namespace {
+
 TVMFFIAny DataflowVarVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const DataflowVarNode* self =
       ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const DataflowVarNode>(value);
@@ -201,6 +342,239 @@ TVMFFIAny DataflowVarMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::details::UnchangedOrUnsafe::MoveToTVMFFIAny(std::move(result));
 }
 
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  DataflowVarNode::RegisterReflection();
+  refl::TypeAttrDef<DataflowVarNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&DataflowVarVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&DataflowVarMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&DataflowVarMaybeInplaceMutate));
+
+  refl::GlobalDef().def("relax.DataflowVar",
+                        [](ffi::String name, ffi::Optional<Type> ty_annotation, Span span) {
+                          return DataflowVar(name, ty_annotation, span);
+                        });
+}
+
+Constant::Constant(runtime::Tensor data, ffi::Optional<Type> ty_annotation, Span span) {
+  ffi::ObjectPtr<ConstantNode> n = ffi::make_object<ConstantNode>();
+  n->data = std::move(data);
+  n->span = std::move(span);
+
+  // set type.
+  ffi::Array<PrimExpr> values;
+  auto shape_tuple = n->data.Shape();
+  for (size_t dim = 0; dim < shape_tuple.size(); ++dim) {
+    values.push_back(IntImm::Int64(shape_tuple[dim]));
+  }
+  if (ty_annotation.has_value()) {
+    n->ty = ty_annotation.value();
+  } else {
+    TensorType tinfo(ShapeExpr(values), PrimType(n->data.DataType()), VDevice(), span);
+    n->ty = tinfo;
+  }
+
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  ConstantNode::RegisterReflection();
+  refl::TypeAttrDef<ConstantNode>()
+      .attr(refl::type_attr::kStructuralVisit,
+            reinterpret_cast<void*>(&TypeOnlyExprVisit<ConstantNode>))
+      .attr(refl::type_attr::kStructuralMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMutate<ConstantNode>))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<ConstantNode>));
+
+  refl::GlobalDef().def("relax.Constant",
+                        [](runtime::Tensor data, ffi::Optional<Type> ty_annotation = std::nullopt,
+                           Span span = Span()) { return Constant(data, ty_annotation, span); });
+}
+
+StringImm::StringImm(ffi::String value, Span span) {
+  ffi::ObjectPtr<StringImmNode> n = ffi::make_object<StringImmNode>();
+  n->value = std::move(value);
+  n->span = std::move(span);
+  n->ty = AnyType();
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  StringImmNode::RegisterReflection();
+  refl::TypeAttrDef<StringImmNode>()
+      .attr(refl::type_attr::kStructuralVisit,
+            reinterpret_cast<void*>(&TypeOnlyExprVisit<StringImmNode>))
+      .attr(refl::type_attr::kStructuralMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMutate<StringImmNode>))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<StringImmNode>));
+
+  refl::GlobalDef().def("relax.StringImm",
+                        [](ffi::String value, Span span) { return StringImm(value, span); });
+}
+
+DataTypeImm::DataTypeImm(DLDataType value, Span span) {
+  ffi::ObjectPtr<DataTypeImmNode> n = ffi::make_object<DataTypeImmNode>();
+  n->value = value;
+  n->span = std::move(span);
+  n->ty = AnyType();
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  DataTypeImmNode::RegisterReflection();
+  refl::TypeAttrDef<DataTypeImmNode>()
+      .attr(refl::type_attr::kStructuralVisit,
+            reinterpret_cast<void*>(&TypeOnlyExprVisit<DataTypeImmNode>))
+      .attr(refl::type_attr::kStructuralMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMutate<DataTypeImmNode>))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<DataTypeImmNode>));
+
+  refl::GlobalDef().def("relax.DataTypeImm",
+                        [](DLDataType value, Span span) { return DataTypeImm(value, span); });
+}
+
+MatchCast::MatchCast(Var var, Expr value, Type ty, Span span) {
+  ffi::ObjectPtr<MatchCastNode> n = ffi::make_object<MatchCastNode>();
+  TVM_FFI_ICHECK(var.defined()) << "MatchCast requires var to be defined";
+  n->var = std::move(var);
+  n->value = std::move(value);
+  n->ty = std::move(ty);
+  n->span = span;
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  MatchCastNode::RegisterReflection();
+
+  refl::GlobalDef().def("relax.MatchCast", [](Var var, Expr value, Type ty, Span span) {
+    return MatchCast(var, value, ty, span);
+  });
+}
+
+VarBinding::VarBinding(Var var, Expr value, Span span) {
+  ffi::ObjectPtr<VarBindingNode> n = ffi::make_object<VarBindingNode>();
+  n->var = std::move(var);
+  n->value = std::move(value);
+  n->span = span;
+  data_ = std::move(n);
+}
+
+bool VarBindingNode::SEqual(const VarBindingNode* other,
+                            ffi::TypedFunction<bool(AnyView, AnyView, bool, AnyView)> equal) const {
+  if (value->IsInstance<FunctionNode>()) {
+    // Recursive function definitions may reference the bound variable
+    // within the value being bound.  In these cases, the
+    // var comparison must occur first to define the var, to ensure it is
+    // defined at point of use.
+    return equal(var, other->var, true, "var") && equal(value, other->value, false, "value");
+  } else {
+    // In all other cases, visit the bound value before the variable
+    // it is bound to, in order to provide better error messages.
+    return equal(value, other->value, false, "value") && equal(var, other->var, true, "var");
+  }
+}
+
+int64_t VarBindingNode::SHash(int64_t init_hash,
+                              ffi::TypedFunction<int64_t(AnyView, int64_t, bool)> hash) const {
+  int64_t hash_value = init_hash;
+  if (value->IsInstance<FunctionNode>()) {
+    hash_value = hash(var, hash_value, true);
+    hash_value = hash(value, hash_value, false);
+  } else {
+    hash_value = hash(value, hash_value, false);
+    hash_value = hash(var, hash_value, true);
+  }
+  return hash_value;
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  VarBindingNode::RegisterReflection();
+
+  refl::GlobalDef().def("relax.VarBinding", [](Var var, Expr value, Span span) {
+    return VarBinding(var, value, span);
+  });
+}
+
+BindingBlock::BindingBlock(ffi::Array<Binding> bindings, Span span) {
+  ffi::ObjectPtr<BindingBlockNode> n = ffi::make_object<BindingBlockNode>();
+  n->bindings = std::move(bindings);
+  n->span = span;
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  BindingBlockNode::RegisterReflection();
+
+  refl::GlobalDef().def("relax.BindingBlock", [](ffi::Array<Binding> bindings, Span span) {
+    return BindingBlock(bindings, span);
+  });
+}
+
+BindingBlockNode* BindingBlock::CopyOnWrite() {
+  // The `TVM_DEFINE_OBJECT_REF_COW_METHOD` cannot be used for
+  // BindingBlock, because it is the base class for `DataflowBlock`.
+  // If the `TVM_DEFINE_OBJECT_REF_COW_METHOD` were used, the
+  // automatic implementation would erroneously convert from a
+  // `DataflowBlock` to a `BindingBlock`.
+  TVM_FFI_ICHECK(data_ != nullptr);
+  if (!data_.unique()) {
+    ffi::ObjectPtr<BindingBlockNode> node;
+    if (auto dataflow_block = as<DataflowBlockNode>()) {
+      node = ffi::make_object<DataflowBlockNode>(*dataflow_block);
+    } else {
+      node = ffi::make_object<BindingBlockNode>(*(operator->()));
+    }
+    ffi::ObjectPtr<ffi::Object>(std::move(node)).swap(data_);
+  }
+  return static_cast<BindingBlockNode*>(data_.get());
+}
+
+DataflowBlock::DataflowBlock(ffi::Array<Binding> bindings, Span span) {
+  ffi::ObjectPtr<DataflowBlockNode> n = ffi::make_object<DataflowBlockNode>();
+  n->bindings = std::move(bindings);
+  n->span = span;
+  data_ = std::move(n);
+}
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  DataflowBlockNode::RegisterReflection();
+
+  refl::GlobalDef().def("relax.DataflowBlock", [](ffi::Array<Binding> bindings, Span span) {
+    return DataflowBlock(bindings, span);
+  });
+}
+
+SeqExpr::SeqExpr(Expr body) {
+  if (auto seq = body.as<SeqExpr>()) {
+    *this = seq.value();
+  } else {
+    *this = SeqExpr(ffi::Array<BindingBlock>{}, body);
+  }
+}
+
+SeqExpr::SeqExpr(ffi::Array<BindingBlock> blocks, Expr body, Span span) {
+  ffi::ObjectPtr<SeqExprNode> n = ffi::make_object<SeqExprNode>();
+  n->blocks = std::move(blocks);
+  n->body = std::move(body);
+  n->span = span;
+  data_ = std::move(n);
+}
+
+namespace {
+
 TVMFFIAny SeqExprVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   const SeqExprNode* self =
       ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const SeqExprNode>(value);
@@ -250,68 +624,85 @@ TVMFFIAny SeqExprMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
   return ffi::Unchanged().CopyToTVMFFIAny();
 }
 
-TVMFFIAny IfVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
-  const IfNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value);
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->ty));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->cond));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->true_branch));
-  TVM_FFI_S_VISIT_MAYBE_EARLY_RETURN(visitor->VisitExpected(self->false_branch));
-  return ffi::AnyView(nullptr).CopyToTVMFFIAny();
+}  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  SeqExprNode::RegisterReflection();
+  refl::TypeAttrDef<SeqExprNode>()
+      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&SeqExprVisit))
+      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&SeqExprMutate))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&SeqExprMaybeInplaceMutate));
+
+  refl::GlobalDef().def("relax.SeqExpr", [](ffi::Array<BindingBlock> blocks, Expr body, Span span) {
+    return SeqExpr(blocks, body, span);
+  });
 }
 
-TVMFFIAny IfMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  const IfNode* self =
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value);
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
-                                    mutator->MutateExpected(self->ty));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, mapped_cond,
-                                    mutator->MutateExpected(self->cond));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_true_branch,
-                                    mutator->MutateExpected(self->true_branch));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_false_branch,
-                                    mutator->MutateExpected(self->false_branch));
-  if (mapped_ty.UnchangedOrSameAs(self->ty) && mapped_cond.UnchangedOrSameAs(self->cond) &&
-      mapped_true_branch.UnchangedOrSameAs(self->true_branch) &&
-      mapped_false_branch.UnchangedOrSameAs(self->false_branch)) {
-    return ffi::Unchanged().CopyToTVMFFIAny();
+Function::Function(ffi::Array<Var> params, Expr body, ffi::Optional<Type> ret_ty, bool is_pure,
+                   DictAttrs attrs, Span span) {
+  // Set the function type.
+  // For function, we take a conservative approach and require the function type
+  // to be known at construction time.
+  ffi::Array<Type> param_ty;
+
+  for (const Var& param : params) {
+    TVM_FFI_ICHECK(!param->ty.IsMissing()) << "relax.Function requires params to contain ty";
+    param_ty.push_back(GetType(param));
   }
-  ffi::ObjectPtr<IfNode> copy = ffi::make_object<IfNode>(*self);
-  copy->ty = std::move(mapped_ty).ValueOrUnchanged(std::move(copy->ty));
-  copy->cond = std::move(mapped_cond).ValueOrUnchanged(std::move(copy->cond));
-  copy->true_branch = std::move(mapped_true_branch).ValueOrUnchanged(std::move(copy->true_branch));
-  copy->false_branch =
-      std::move(mapped_false_branch).ValueOrUnchanged(std::move(copy->false_branch));
-  return ffi::details::AnyUnsafe::MoveAnyToTVMFFIAny(ffi::Any(std::move(copy)));
+
+  ffi::Optional<Type> body_ty;
+
+  if (!body->ty.IsMissing()) {
+    body_ty = GetType(body);
+  }
+
+  TVM_FFI_ICHECK(body_ty.has_value() || ret_ty.has_value())
+      << "Function must be constructed with either "
+      << "an explicit type for the return type, "
+      << "or a normalized body with type.";
+
+  // Use the body's type if there is no explicit return type,
+  // or if the body may provide a more granular return type.
+  bool use_body_ty =
+      !ret_ty.has_value() || (body_ty && ret_ty && IsBaseOf(ret_ty.value(), body_ty.value()));
+
+  if (use_body_ty) {
+    // MatchCast nodes within the body may introduce new symbolic
+    // variables.  These are in-scope for the function body, but not
+    // for the function's return type.  When hoisting the body's type
+    // to the function return type, symbolic variables may only be
+    // used if they were defined by the function's parameters.
+    auto f_var_map = [&] {
+      auto tir_vars = DefinableTIRVarsInType(TupleType(params.Map(GetType)));
+      std::unordered_set<tirx::Var> lookup(tir_vars.begin(), tir_vars.end());
+      return [lookup = std::move(lookup)](const Var& var) -> ffi::Optional<Expr> {
+        if (auto prim_var = var.as<tirx::PrimVar>(); prim_var && lookup.count(prim_var.value())) {
+          return prim_var.value().as_or_throw<PrimExpr>();
+        }
+        return std::nullopt;
+      };
+    }();
+    ret_ty = EraseToWellDefined(body_ty.value(), f_var_map);
+  }
+
+  FuncType func_ty(param_ty, ret_ty.value(), is_pure);
+
+  // set the fields
+  ffi::ObjectPtr<FunctionNode> n = ffi::make_object<FunctionNode>();
+  n->params = std::move(params);
+  n->body = std::move(body);
+  n->ret_ty = ret_ty.value();
+  n->is_pure = is_pure;
+  n->ty = std::move(func_ty);
+  n->attrs = std::move(attrs);
+  n->span = std::move(span);
+  data_ = std::move(n);
 }
 
-TVMFFIAny IfMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator, ffi::AnyView value) noexcept {
-  IfNode* self = const_cast<IfNode*>(
-      ffi::details::AnyUnsafe::RawObjectPtrFromAnyViewAfterCheck<const IfNode>(value));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Type>, mapped_ty,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->ty));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<Expr>, mapped_cond,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->cond));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(ffi::UnchangedOr<SeqExpr>, mapped_true_branch,
-                                    mutator->MaybeInplaceMutateIfUniqueExpected(self->true_branch));
-  TVM_FFI_S_MUTATE_ASSIGN_OR_RETURN(
-      ffi::UnchangedOr<SeqExpr>, mapped_false_branch,
-      mutator->MaybeInplaceMutateIfUniqueExpected(self->false_branch));
-  if (!mapped_ty.UnchangedOrSameAs(self->ty)) self->ty = std::move(mapped_ty).ValueUnchecked();
-  if (!mapped_cond.UnchangedOrSameAs(self->cond)) {
-    self->cond = std::move(mapped_cond).ValueUnchecked();
-  }
-  if (!mapped_true_branch.UnchangedOrSameAs(self->true_branch)) {
-    self->true_branch = std::move(mapped_true_branch).ValueUnchecked();
-  }
-  if (!mapped_false_branch.UnchangedOrSameAs(self->false_branch)) {
-    self->false_branch = std::move(mapped_false_branch).ValueUnchecked();
-  }
-  return ffi::Unchanged().CopyToTVMFFIAny();
-}
+namespace {
 
-// Parameters precede the reflected ty field in this hook triple so their pattern region establishes
-// the remap before the derived function type can refer to those symbolic definitions.
 TVMFFIAny FunctionVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
   // skips: attrs (metadata), is_pure (scalar)
   const FunctionNode* self =
@@ -383,381 +774,13 @@ TVMFFIAny FunctionMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  ShapeExprNode::RegisterReflection();
-  BindingNode::RegisterReflection();
-  DataflowVarNode::RegisterReflection();
-  ConstantNode::RegisterReflection();
-  StringImmNode::RegisterReflection();
-  DataTypeImmNode::RegisterReflection();
-  MatchCastNode::RegisterReflection();
-  VarBindingNode::RegisterReflection();
-  BindingBlockNode::RegisterReflection();
-  DataflowBlockNode::RegisterReflection();
-  SeqExprNode::RegisterReflection();
-  IfNode::RegisterReflection();
   FunctionNode::RegisterReflection();
-  ExternFuncNode::RegisterReflection();
-  refl::TypeAttrDef<ShapeExprNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&ShapeExprVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&ShapeExprMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&ShapeExprMaybeInplaceMutate));
-  refl::TypeAttrDef<DataflowVarNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&DataflowVarVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&DataflowVarMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&DataflowVarMaybeInplaceMutate));
-  refl::TypeAttrDef<ConstantNode>()
-      .attr(refl::type_attr::kStructuralVisit,
-            reinterpret_cast<void*>(&TypeOnlyExprVisit<ConstantNode>))
-      .attr(refl::type_attr::kStructuralMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMutate<ConstantNode>))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<ConstantNode>));
-  refl::TypeAttrDef<StringImmNode>()
-      .attr(refl::type_attr::kStructuralVisit,
-            reinterpret_cast<void*>(&TypeOnlyExprVisit<StringImmNode>))
-      .attr(refl::type_attr::kStructuralMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMutate<StringImmNode>))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<StringImmNode>));
-  refl::TypeAttrDef<DataTypeImmNode>()
-      .attr(refl::type_attr::kStructuralVisit,
-            reinterpret_cast<void*>(&TypeOnlyExprVisit<DataTypeImmNode>))
-      .attr(refl::type_attr::kStructuralMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMutate<DataTypeImmNode>))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<DataTypeImmNode>));
-  refl::TypeAttrDef<SeqExprNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&SeqExprVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&SeqExprMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&SeqExprMaybeInplaceMutate));
-  refl::TypeAttrDef<IfNode>()
-      .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&IfVisit))
-      .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&IfMutate))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&IfMaybeInplaceMutate));
   refl::TypeAttrDef<FunctionNode>()
       .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&FunctionVisit))
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&FunctionMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&FunctionMaybeInplaceMutate));
-  refl::TypeAttrDef<ExternFuncNode>()
-      .attr(refl::type_attr::kStructuralVisit,
-            reinterpret_cast<void*>(&TypeOnlyExprVisit<ExternFuncNode>))
-      .attr(refl::type_attr::kStructuralMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMutate<ExternFuncNode>))
-      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
-            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<ExternFuncNode>));
-}
 
-If::If(Expr cond, Expr true_branch, Expr false_branch, Span span) {
-  ffi::ObjectPtr<IfNode> n = ffi::make_object<IfNode>();
-  n->cond = std::move(cond);
-  n->true_branch = std::move(true_branch);
-  n->false_branch = std::move(false_branch);
-  n->span = std::move(span);
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.If", [](Expr cond, Expr true_branch, Expr false_branch, Span span) {
-    return If(cond, true_branch, false_branch, span);
-  });
-}
-
-ShapeExpr::ShapeExpr(ffi::Array<PrimExpr> values, Span span) {
-  ffi::ObjectPtr<ShapeExprNode> n = ffi::make_object<ShapeExprNode>();
-
-  n->values = values.Map([](PrimExpr value) {
-    if (value->IsInstance<IntImmNode>()) {
-      return tvm::cast(PrimType::Int(64), value);
-    }
-    TVM_FFI_ICHECK(value.ty().MatchesElementType(DLDataTypeCode::kDLInt, 64))
-        << "the value in ShapeType can only have dtype of int64";
-    return value;
-  });
-  n->span = span;
-  n->ty = ShapeType(values, span);
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.ShapeExpr", [](ffi::Array<PrimExpr> values, Span span) {
-    return ShapeExpr(values, span);
-  });
-}
-
-DataflowVar::DataflowVar(ffi::String name, ffi::Optional<Type> ty_annotation, Span span) {
-  ffi::ObjectPtr<DataflowVarNode> n = ffi::make_object<DataflowVarNode>();
-  n->name = std::move(name);
-  if (ty_annotation.has_value()) {
-    n->ty = ty_annotation.value();
-  }
-  n->span = std::move(span);
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.DataflowVar",
-                        [](ffi::String name, ffi::Optional<Type> ty_annotation, Span span) {
-                          return DataflowVar(name, ty_annotation, span);
-                        });
-}
-
-Constant::Constant(runtime::Tensor data, ffi::Optional<Type> ty_annotation, Span span) {
-  ffi::ObjectPtr<ConstantNode> n = ffi::make_object<ConstantNode>();
-  n->data = std::move(data);
-  n->span = std::move(span);
-
-  // set type.
-  ffi::Array<PrimExpr> values;
-  auto shape_tuple = n->data.Shape();
-  for (size_t dim = 0; dim < shape_tuple.size(); ++dim) {
-    values.push_back(IntImm::Int64(shape_tuple[dim]));
-  }
-  if (ty_annotation.has_value()) {
-    n->ty = ty_annotation.value();
-  } else {
-    TensorType tinfo(ShapeExpr(values), PrimType(n->data.DataType()), VDevice(), span);
-    n->ty = tinfo;
-  }
-
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.Constant",
-                        [](runtime::Tensor data, ffi::Optional<Type> ty_annotation = std::nullopt,
-                           Span span = Span()) { return Constant(data, ty_annotation, span); });
-}
-
-StringImm::StringImm(ffi::String value, Span span) {
-  ffi::ObjectPtr<StringImmNode> n = ffi::make_object<StringImmNode>();
-  n->value = std::move(value);
-  n->span = std::move(span);
-  n->ty = AnyType();
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.StringImm",
-                        [](ffi::String value, Span span) { return StringImm(value, span); });
-}
-
-DataTypeImm::DataTypeImm(DLDataType value, Span span) {
-  ffi::ObjectPtr<DataTypeImmNode> n = ffi::make_object<DataTypeImmNode>();
-  n->value = value;
-  n->span = std::move(span);
-  n->ty = AnyType();
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.DataTypeImm",
-                        [](DLDataType value, Span span) { return DataTypeImm(value, span); });
-}
-
-MatchCast::MatchCast(Var var, Expr value, Type ty, Span span) {
-  ffi::ObjectPtr<MatchCastNode> n = ffi::make_object<MatchCastNode>();
-  TVM_FFI_ICHECK(var.defined()) << "MatchCast requires var to be defined";
-  n->var = std::move(var);
-  n->value = std::move(value);
-  n->ty = std::move(ty);
-  n->span = span;
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.MatchCast", [](Var var, Expr value, Type ty, Span span) {
-    return MatchCast(var, value, ty, span);
-  });
-}
-
-VarBinding::VarBinding(Var var, Expr value, Span span) {
-  ffi::ObjectPtr<VarBindingNode> n = ffi::make_object<VarBindingNode>();
-  n->var = std::move(var);
-  n->value = std::move(value);
-  n->span = span;
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.VarBinding", [](Var var, Expr value, Span span) {
-    return VarBinding(var, value, span);
-  });
-}
-
-bool VarBindingNode::SEqual(const VarBindingNode* other,
-                            ffi::TypedFunction<bool(AnyView, AnyView, bool, AnyView)> equal) const {
-  if (value->IsInstance<FunctionNode>()) {
-    // Recursive function definitions may reference the bound variable
-    // within the value being bound.  In these cases, the
-    // var comparison must occur first to define the var, to ensure it is
-    // defined at point of use.
-    return equal(var, other->var, true, "var") && equal(value, other->value, false, "value");
-  } else {
-    // In all other cases, visit the bound value before the variable
-    // it is bound to, in order to provide better error messages.
-    return equal(value, other->value, false, "value") && equal(var, other->var, true, "var");
-  }
-}
-
-int64_t VarBindingNode::SHash(int64_t init_hash,
-                              ffi::TypedFunction<int64_t(AnyView, int64_t, bool)> hash) const {
-  int64_t hash_value = init_hash;
-  if (value->IsInstance<FunctionNode>()) {
-    hash_value = hash(var, hash_value, true);
-    hash_value = hash(value, hash_value, false);
-  } else {
-    hash_value = hash(value, hash_value, false);
-    hash_value = hash(var, hash_value, true);
-  }
-  return hash_value;
-}
-
-BindingBlock::BindingBlock(ffi::Array<Binding> bindings, Span span) {
-  ffi::ObjectPtr<BindingBlockNode> n = ffi::make_object<BindingBlockNode>();
-  n->bindings = std::move(bindings);
-  n->span = span;
-  data_ = std::move(n);
-}
-
-BindingBlockNode* BindingBlock::CopyOnWrite() {
-  // The `TVM_DEFINE_OBJECT_REF_COW_METHOD` cannot be used for
-  // BindingBlock, because it is the base class for `DataflowBlock`.
-  // If the `TVM_DEFINE_OBJECT_REF_COW_METHOD` were used, the
-  // automatic implementation would erroneously convert from a
-  // `DataflowBlock` to a `BindingBlock`.
-  TVM_FFI_ICHECK(data_ != nullptr);
-  if (!data_.unique()) {
-    ffi::ObjectPtr<BindingBlockNode> node;
-    if (auto dataflow_block = as<DataflowBlockNode>()) {
-      node = ffi::make_object<DataflowBlockNode>(*dataflow_block);
-    } else {
-      node = ffi::make_object<BindingBlockNode>(*(operator->()));
-    }
-    ffi::ObjectPtr<ffi::Object>(std::move(node)).swap(data_);
-  }
-  return static_cast<BindingBlockNode*>(data_.get());
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.BindingBlock", [](ffi::Array<Binding> bindings, Span span) {
-    return BindingBlock(bindings, span);
-  });
-}
-
-DataflowBlock::DataflowBlock(ffi::Array<Binding> bindings, Span span) {
-  ffi::ObjectPtr<DataflowBlockNode> n = ffi::make_object<DataflowBlockNode>();
-  n->bindings = std::move(bindings);
-  n->span = span;
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.DataflowBlock", [](ffi::Array<Binding> bindings, Span span) {
-    return DataflowBlock(bindings, span);
-  });
-}
-
-SeqExpr::SeqExpr(Expr body) {
-  if (auto seq = body.as<SeqExpr>()) {
-    *this = seq.value();
-  } else {
-    *this = SeqExpr(ffi::Array<BindingBlock>{}, body);
-  }
-}
-
-SeqExpr::SeqExpr(ffi::Array<BindingBlock> blocks, Expr body, Span span) {
-  ffi::ObjectPtr<SeqExprNode> n = ffi::make_object<SeqExprNode>();
-  n->blocks = std::move(blocks);
-  n->body = std::move(body);
-  n->span = span;
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef().def("relax.SeqExpr", [](ffi::Array<BindingBlock> blocks, Expr body, Span span) {
-    return SeqExpr(blocks, body, span);
-  });
-}
-
-Function::Function(ffi::Array<Var> params, Expr body, ffi::Optional<Type> ret_ty, bool is_pure,
-                   DictAttrs attrs, Span span) {
-  // Set the function type.
-  // For function, we take a conservative approach and require the function type
-  // to be known at construction time.
-  ffi::Array<Type> param_ty;
-
-  for (const Var& param : params) {
-    TVM_FFI_ICHECK(!param->ty.IsMissing()) << "relax.Function requires params to contain ty";
-    param_ty.push_back(GetType(param));
-  }
-
-  ffi::Optional<Type> body_ty;
-
-  if (!body->ty.IsMissing()) {
-    body_ty = GetType(body);
-  }
-
-  TVM_FFI_ICHECK(body_ty.has_value() || ret_ty.has_value())
-      << "Function must be constructed with either "
-      << "an explicit type for the return type, "
-      << "or a normalized body with type.";
-
-  // Use the body's type if there is no explicit return type,
-  // or if the body may provide a more granular return type.
-  bool use_body_ty =
-      !ret_ty.has_value() || (body_ty && ret_ty && IsBaseOf(ret_ty.value(), body_ty.value()));
-
-  if (use_body_ty) {
-    // MatchCast nodes within the body may introduce new symbolic
-    // variables.  These are in-scope for the function body, but not
-    // for the function's return type.  When hoisting the body's type
-    // to the function return type, symbolic variables may only be
-    // used if they were defined by the function's parameters.
-    auto f_var_map = [&] {
-      auto tir_vars = DefinableTIRVarsInType(TupleType(params.Map(GetType)));
-      std::unordered_set<tirx::Var> lookup(tir_vars.begin(), tir_vars.end());
-      return [lookup = std::move(lookup)](const Var& var) -> ffi::Optional<Expr> {
-        if (auto prim_var = var.as<tirx::PrimVar>(); prim_var && lookup.count(prim_var.value())) {
-          return prim_var.value().as_or_throw<PrimExpr>();
-        }
-        return std::nullopt;
-      };
-    }();
-    ret_ty = EraseToWellDefined(body_ty.value(), f_var_map);
-  }
-
-  FuncType func_ty(param_ty, ret_ty.value(), is_pure);
-
-  // set the fields
-  ffi::ObjectPtr<FunctionNode> n = ffi::make_object<FunctionNode>();
-  n->params = std::move(params);
-  n->body = std::move(body);
-  n->ret_ty = ret_ty.value();
-  n->is_pure = is_pure;
-  n->ty = std::move(func_ty);
-  n->attrs = std::move(attrs);
-  n->span = std::move(span);
-  data_ = std::move(n);
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
   refl::GlobalDef().def("relax.Function",
                         [](ffi::Array<Var> params, Expr body, ffi::Optional<Type> ret_ty,
                            bool is_pure, DictAttrs attrs, Span span) {
@@ -845,6 +868,15 @@ ExternFunc::ExternFunc(ffi::String global_symbol, Type ty, Span span) {
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
+  ExternFuncNode::RegisterReflection();
+  refl::TypeAttrDef<ExternFuncNode>()
+      .attr(refl::type_attr::kStructuralVisit,
+            reinterpret_cast<void*>(&TypeOnlyExprVisit<ExternFuncNode>))
+      .attr(refl::type_attr::kStructuralMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMutate<ExternFuncNode>))
+      .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
+            reinterpret_cast<void*>(&TypeOnlyExprMaybeInplaceMutate<ExternFuncNode>));
+
   refl::GlobalDef().def("relax.ExternFunc",
                         [](ffi::String global_symbol, ffi::Optional<Type> ty, Span span) {
                           if (ty.has_value()) {
