+++ src/s_tir/data_layout.cc
@@ -43,11 +43,6 @@ using tirx::IterVar;
 using tirx::IterVarNode;
 using tirx::Var;
 
-TVM_FFI_STATIC_INIT_BLOCK() {
-  SLayoutNode::RegisterReflection();
-  SBijectiveLayoutNode::RegisterReflection();
-}
-
 const SLayoutAxis SLayoutAxis::UPPER_CASE[] = {
     SLayoutAxis('A'), SLayoutAxis('B'), SLayoutAxis('C'), SLayoutAxis('D'), SLayoutAxis('E'),
     SLayoutAxis('F'), SLayoutAxis('G'), SLayoutAxis('H'), SLayoutAxis('I'), SLayoutAxis('J'),
@@ -216,6 +211,18 @@ SLayout::SLayout(const std::string& name, PrimType index_ty) {  // NOLINT(*)
   data_ = std::move(node);
 }
 
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  SLayoutNode::RegisterReflection();
+  refl::TypeAttrDef<SLayoutNode>().def(refl::type_attr::kRepr,
+                                       [](SLayout l, ffi::Function) -> ffi::String {
+                                         return "SLayout(" + std::string(l->name) + ")";
+                                       });
+
+  refl::GlobalDef().def("s_tir.SLayout",
+                        [](std::string name, PrimType dtype) { return SLayout(name, dtype); });
+}
+
 SLayout SLayout::SubLayout(size_t pos, size_t len) const {
   if (!defined() || pos > ndim()) return SLayout::Undef();
   if (len == 0) return SLayout(ffi::Array<IterVar>());
@@ -288,14 +295,6 @@ int32_t SLayout::FactorOf(const SLayoutAxis& axis) const {
   factor = has_sub ? factor : -1;
 
   return factor;
-}
-
-TVM_FFI_STATIC_INIT_BLOCK() {
-  namespace refl = tvm::ffi::reflection;
-  refl::TypeAttrDef<SLayoutNode>().def(refl::type_attr::kRepr,
-                                       [](SLayout l, ffi::Function) -> ffi::String {
-                                         return "SLayout(" + std::string(l->name) + ")";
-                                       });
 }
 
 inline bool GetStoreRule(ffi::Array<PrimExpr>* index_rule, ffi::Array<PrimExpr>* shape_rule,
@@ -589,17 +588,22 @@ SBijectiveLayout::SBijectiveLayout(SLayout src_layout, SLayout dst_layout) {
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
+  SBijectiveLayoutNode::RegisterReflection();
   refl::TypeAttrDef<SBijectiveLayoutNode>().def(
       refl::type_attr::kRepr, [](SBijectiveLayout bl, ffi::Function) -> ffi::String {
         return "SBijectiveLayout(" + std::string(bl->src_layout.name()) + "->" +
                std::string(bl->dst_layout.name()) + ")";
       });
+
+  refl::GlobalDef().def("s_tir.SBijectiveLayout",
+                        [](SLayout src_layout, SLayout dst_layout) -> SBijectiveLayout {
+                          return SBijectiveLayout(src_layout, dst_layout);
+                        });
 }
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   refl::GlobalDef()
-      .def("s_tir.SLayout", [](std::string name, PrimType dtype) { return SLayout(name, dtype); })
       .def("s_tir.SLayoutIndexOf",
            [](SLayout layout, std::string axis) -> int { return layout.IndexOf(axis); })
       .def("s_tir.SLayoutFactorOf",
@@ -611,10 +615,6 @@ TVM_FFI_STATIC_INIT_BLOCK() {
            [](SLayout layout, int idx) -> std::string {
              const auto& axis = layout.PackedAxisAt(idx);
              return axis->var->name;
-           })
-      .def("s_tir.SBijectiveLayout",
-           [](SLayout src_layout, SLayout dst_layout) -> SBijectiveLayout {
-             return SBijectiveLayout(src_layout, dst_layout);
            })
       .def_method("s_tir.SBijectiveLayoutForwardIndex", &SBijectiveLayout::ForwardIndex)
       .def_method("s_tir.SBijectiveLayoutBackwardIndex", &SBijectiveLayout::BackwardIndex)
