+++ src/s_tir/schedule/instruction.cc
@@ -27,10 +27,7 @@ namespace s_tir {
 using namespace tvm::prim;
 using namespace tvm::tirx;
 
-TVM_FFI_STATIC_INIT_BLOCK() {
-  InstructionKindNode::RegisterReflection();
-  InstructionNode::RegisterReflection();
-}
+TVM_FFI_STATIC_INIT_BLOCK() { InstructionKindNode::RegisterReflection(); }
 
 bool InstructionKindNode::IsPostproc() const {
   static InstructionKind inst_enter_postproc = InstructionKind::Get("EnterPostproc");
@@ -47,26 +44,8 @@ Instruction::Instruction(InstructionKind kind, ffi::Array<Any> inputs, ffi::Arra
   this->data_ = std::move(n);
 }
 
-using InstructionKindRegistry = AttrRegistry<InstructionKindRegEntry, InstructionKind>;
-
-InstructionKind InstructionKind::Get(const ffi::String& name) {
-  const InstructionKindRegEntry* reg = InstructionKindRegistry::Global()->Get(name);
-  TVM_FFI_CHECK(reg != nullptr, AttributeError)
-      << "Instruction kind " << name << " is not registered";
-  return reg->inst_kind_;
-}
-
-InstructionKindRegEntry::InstructionKindRegEntry(uint32_t reg_index) {
-  this->inst_kind_ = InstructionKind(ffi::make_object<InstructionKindNode>());
-}
-
-InstructionKindRegEntry& InstructionKindRegEntry::RegisterOrGet(const ffi::String& name) {
-  return InstructionKindRegistry::Global()->RegisterOrGet(name);
-}
-
-/**************** Repr ****************/
-
 namespace {
+
 ffi::String InstructionAsPythonRepr(const InstructionNode* self) {
   ffi::Array<Any> inputs;
   inputs.reserve(self->inputs.size());
@@ -103,7 +82,42 @@ ffi::String InstructionAsPythonRepr(const InstructionNode* self) {
       /*decision=*/Any(nullptr),
       /*outputs=*/ffi::Array<ffi::String>(self->outputs.size(), ffi::String("_")));
 }
+
 }  // namespace
+
+TVM_FFI_STATIC_INIT_BLOCK() {
+  namespace refl = tvm::ffi::reflection;
+  InstructionNode::RegisterReflection();
+  refl::TypeAttrDef<InstructionNode>().def(refl::type_attr::kRepr,
+                                           [](Instruction inst, ffi::Function) -> ffi::String {
+                                             return InstructionAsPythonRepr(inst.get());
+                                           });
+
+  refl::GlobalDef().def("s_tir.schedule.Instruction",
+                        [](InstructionKind kind, ffi::Array<Any> inputs, ffi::Array<Any> attrs,
+                           ffi::Array<Any> outputs) -> Instruction {
+                          return Instruction(kind, inputs, attrs, outputs);
+                        });
+}
+
+using InstructionKindRegistry = AttrRegistry<InstructionKindRegEntry, InstructionKind>;
+
+InstructionKind InstructionKind::Get(const ffi::String& name) {
+  const InstructionKindRegEntry* reg = InstructionKindRegistry::Global()->Get(name);
+  TVM_FFI_CHECK(reg != nullptr, AttributeError)
+      << "Instruction kind " << name << " is not registered";
+  return reg->inst_kind_;
+}
+
+InstructionKindRegEntry::InstructionKindRegEntry(uint32_t reg_index) {
+  this->inst_kind_ = InstructionKind(ffi::make_object<InstructionKindNode>());
+}
+
+InstructionKindRegEntry& InstructionKindRegEntry::RegisterOrGet(const ffi::String& name) {
+  return InstructionKindRegistry::Global()->RegisterOrGet(name);
+}
+
+/**************** Repr ****************/
 
 // AC: kRepr already registered below in TVM_FFI_STATIC_INIT_BLOCK.
 
@@ -111,17 +125,7 @@ ffi::String InstructionAsPythonRepr(const InstructionNode* self) {
 
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
-  refl::GlobalDef()
-      .def("s_tir.schedule.InstructionKindGet", InstructionKind::Get)
-      .def("s_tir.schedule.Instruction",
-           [](InstructionKind kind, ffi::Array<Any> inputs, ffi::Array<Any> attrs,
-              ffi::Array<Any> outputs) -> Instruction {
-             return Instruction(kind, inputs, attrs, outputs);
-           });
-  refl::TypeAttrDef<InstructionNode>().def(refl::type_attr::kRepr,
-                                           [](Instruction inst, ffi::Function) -> ffi::String {
-                                             return InstructionAsPythonRepr(inst.get());
-                                           });
+  refl::GlobalDef().def("s_tir.schedule.InstructionKindGet", InstructionKind::Get);
 }
 
 }  // namespace s_tir
