+++ src/tirx/ir/function.cc
@@ -34,6 +34,68 @@
 namespace tvm {
 namespace tirx {
 
+TVM_FFI_STATIC_INIT_BLOCK() { TensorIntrinNode::RegisterReflection(); }
+
+namespace {
+tvm::Type InferType(const PrimFunc& prim_func) {
+  ffi::Array<tvm::Type> params;
+  for (const auto& param : prim_func->params) {
+    tvm::Type param_ty = [&]() -> tvm::Type {
+      if (param->ty.as<BufferTypeNode>()) {
+        BufferVar buf(param);
+        relax::ShapeExpr shape(
+            buf->shape.Map([](PrimExpr dim) { return cast(PrimType::Int(64), dim); }));
+        return relax::TensorType(shape, buf->dtype);
+      }
+
+      // A pointer parameter without a buffer annotation is an opaque runtime
+      // object from Relax's perspective (for example, a DLTensor*).  Keep the
+      // same Relax-facing wildcard semantics that opaque handle parameters had
+      // before pointers became exact IR types.
+      if (param->ty.as<PointerTypeNode>()) {
+        return relax::AnyType();
+      }
+
+      return param->ty;
+    }();
+    params.push_back(param_ty);
+  }
+
+  tvm::Type ret = [&]() -> tvm::Type {
+    if (const auto* prim = prim_func->ret_type.as<PrimTypeNode>()) {
+      return tvm::PrimType(prim->dtype);
+    } else if (IsVoidType(prim_func->ret_type)) {
+      return relax::TupleType(ffi::Array<tvm::Type>{});
+    } else {
+      return relax::AnyType();
+    }
+  }();
+
+  bool purity = prim_func->body.defined() ? s_tir::IsPureFunction(prim_func) : false;
+
+  return relax::FuncType(params, ret, purity);
+}
+}  // namespace
+
+// Get the function type of a PrimFunc
+PrimFunc::PrimFunc(ffi::Array<tirx::Var> params, Stmt body, Type ret_type, DictAttrs attrs,
+                   Span span) {
+  if (ret_type.IsMissing()) {
+    ret_type = VoidType();
+  }
+
+  auto n = ffi::make_object<PrimFuncNode>();
+  n->params = std::move(params);
+  n->body = std::move(body);
+  n->ret_type = std::move(ret_type);
+  n->attrs = std::move(attrs);
+  n->ty = relax::FuncType::OpaqueFunc();
+  n->span = std::move(span);
+  data_ = std::move(n);
+
+  (*this)->ty = InferType(*this);
+}
+
 namespace {
 
 TVMFFIAny PrimFuncVisit(ffi::StructuralVisitorObj* visitor, ffi::AnyView value) noexcept {
@@ -102,72 +164,15 @@ TVMFFIAny PrimFuncMaybeInplaceMutate(ffi::StructuralMutatorObj* mutator,
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   PrimFuncNode::RegisterReflection();
-  TensorIntrinNode::RegisterReflection();
   refl::TypeAttrDef<PrimFuncNode>()
       .attr(refl::type_attr::kStructuralVisit, reinterpret_cast<void*>(&PrimFuncVisit))
       .attr(refl::type_attr::kStructuralMutate, reinterpret_cast<void*>(&PrimFuncMutate))
       .attr(refl::type_attr::kStructuralMaybeInplaceMutate,
             reinterpret_cast<void*>(&PrimFuncMaybeInplaceMutate));
-}
 
-namespace {
-tvm::Type InferType(const PrimFunc& prim_func) {
-  ffi::Array<tvm::Type> params;
-  for (const auto& param : prim_func->params) {
-    tvm::Type param_ty = [&]() -> tvm::Type {
-      if (param->ty.as<BufferTypeNode>()) {
-        BufferVar buf(param);
-        relax::ShapeExpr shape(
-            buf->shape.Map([](PrimExpr dim) { return cast(PrimType::Int(64), dim); }));
-        return relax::TensorType(shape, buf->dtype);
-      }
-
-      // A pointer parameter without a buffer annotation is an opaque runtime
-      // object from Relax's perspective (for example, a DLTensor*).  Keep the
-      // same Relax-facing wildcard semantics that opaque handle parameters had
-      // before pointers became exact IR types.
-      if (param->ty.as<PointerTypeNode>()) {
-        return relax::AnyType();
-      }
-
-      return param->ty;
-    }();
-    params.push_back(param_ty);
-  }
-
-  tvm::Type ret = [&]() -> tvm::Type {
-    if (const auto* prim = prim_func->ret_type.as<PrimTypeNode>()) {
-      return tvm::PrimType(prim->dtype);
-    } else if (IsVoidType(prim_func->ret_type)) {
-      return relax::TupleType(ffi::Array<tvm::Type>{});
-    } else {
-      return relax::AnyType();
-    }
-  }();
-
-  bool purity = prim_func->body.defined() ? s_tir::IsPureFunction(prim_func) : false;
-
-  return relax::FuncType(params, ret, purity);
-}
-}  // namespace
-
-// Get the function type of a PrimFunc
-PrimFunc::PrimFunc(ffi::Array<tirx::Var> params, Stmt body, Type ret_type, DictAttrs attrs,
-                   Span span) {
-  if (ret_type.IsMissing()) {
-    ret_type = VoidType();
-  }
-
-  auto n = ffi::make_object<PrimFuncNode>();
-  n->params = std::move(params);
-  n->body = std::move(body);
-  n->ret_type = std::move(ret_type);
-  n->attrs = std::move(attrs);
-  n->ty = relax::FuncType::OpaqueFunc();
-  n->span = std::move(span);
-  data_ = std::move(n);
-
-  (*this)->ty = InferType(*this);
+  refl::GlobalDef().def("tirx.PrimFunc",
+                        [](ffi::Array<tirx::Var> params, Stmt body, Type ret_type, DictAttrs attrs,
+                           Span span) { return PrimFunc(params, body, ret_type, attrs, span); });
 }
 
 FuncType PrimFuncNode::func_type_annotation() const {
@@ -235,10 +240,6 @@ ffi::Optional<TensorIntrin> TensorIntrin::Get(ffi::String name, bool allow_missi
 TVM_FFI_STATIC_INIT_BLOCK() {
   namespace refl = tvm::ffi::reflection;
   refl::GlobalDef()
-      .def("tirx.PrimFunc",
-           [](ffi::Array<tirx::Var> params, Stmt body, Type ret_type, DictAttrs attrs, Span span) {
-             return PrimFunc(params, body, ret_type, attrs, span);
-           })
       .def("tirx.TensorIntrin",
            [](PrimFunc desc_func, PrimFunc intrin_func) {
              return TensorIntrin(desc_func, intrin_func);
